added Google API activity names to logging, requesting permission to write to external storage, change log file name format, various minor changes
