+++ .idea/vcs.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="VcsDirectoryMappings">
+    
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
+
   </component>
 </project>
