+++ app/src/main/java/com/notesapp/ActivityRecognitionService.java
@@ -3,14 +3,9 @@ package com.notesapp;
 import android.app.IntentService;
 import android.content.Intent;
 import android.support.v4.content.LocalBroadcastManager;
-import android.util.Log;
-import android.widget.Toast;
 
 import com.google.android.gms.location.ActivityRecognitionResult;
 import com.google.android.gms.location.DetectedActivity;
-
-import java.util.ArrayList;
-import java.util.List;
 
 
 public class ActivityRecognitionService extends IntentService {
@@ -42,35 +37,5 @@ public class ActivityRecognitionService extends IntentService {
 
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
-
-
-    /*
-    private void sendToMainActivity(ActivityRecognitionResult result) {
-        List<DetectedActivity> probableActivities = result.getProbableActivities();
-        ArrayList<Integer> typeList = new ArrayList<>();
-        ArrayList<Integer> confidenceList = new ArrayList<>();
-
-        for(DetectedActivity activity : probableActivities) {
-            typeList.add(activity.getType());
-            confidenceList.add(activity.getConfidence());
-
-            // logging
-            if(MainActivity.logging) {
-                MainActivity.writeLog(MainActivity.getCurrentTime() +
-                        ": " +
-                        String.valueOf(activity.getType()) +
-                        "(" +
-                        String.valueOf(activity.getConfidence()) +
-                        "%)\n");
-            }
-        }
-
-        Intent intent = new Intent("activityRecognitionIntent");
-        intent.putExtra("type", typeList);
-        intent.putExtra("confidence", confidenceList);
-
-        LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
-    }
-    */
 
 }
