+++ app/src/main/java/com/notesapp/PseudoCluster.java
@@ -7,104 +7,92 @@ import java.util.ArrayList;
 import java.util.List;
 
 
+/**
+ * Represents a single cluster of locations, part of a cluster list in the PseudoClusterList class.
+ */
 public class PseudoCluster {
-    private List<PseudoClusterElement> clusterList;
+    private List<Location> locationCluster;
+    private Location meanLocation;
 
-    private double distanceThreshold;
 
-
-    PseudoCluster(double distanceThreshold) {
-        this.distanceThreshold = distanceThreshold;
-        clusterList = new ArrayList<>();
+    PseudoCluster(Location location){
+        locationCluster = new ArrayList<>();
+        locationCluster.add(location);
+        meanLocation = new Location("");
+        meanLocation.setLatitude(location.getLatitude());
+        meanLocation.setLongitude(location.getLongitude());
     }
 
 
     /**
-     * Counts the total number of elements across all clusters.
-     * @return the number of elements.
-     */
-    int getTotalElementCount() {
-        int size = clusterList.size();
-        if (size == 0) {
-            return 0;
-        }
-        else {
-            int sum = 0;
-            for (PseudoClusterElement cluster : clusterList) {
-                sum += cluster.size();
-            }
-            return sum;
-        }
-    }
-
-
-    /**
-     * Adds a location to the cluster list, into the best fitting cluster.
-     * @param location the Location to be added to a cluster.
+     * Adds a new location to the cluster and calculates a new mean location, based on the new list
+     * of cluster elements.
+     * @param location the Location to be added to the cluster.
      */
     void add(Location location) {
-        int clusterCount = clusterList.size();
-
-        // first location - create first cluster and add the location to it
-        if(clusterCount == 0) {
-            addCluster(location);
-        }
-
-        else {
-
-            // try to place the location in an existing cluster
-            for(int i = 0; i < clusterCount; i++) {
-                // if distance to the mean of i-th cluster < threshold -> place the location
-                // in that cluster
-                if(location.distanceTo(clusterList.get(i).getMeanLocation()) < distanceThreshold) {
-                    clusterList.get(i).add(location);
-                    return;
-                }
-
-            }
-
-            // if not placed -> requires a new cluster
-            addCluster(location);
-
-        }
-
+        locationCluster.add(location);
+        meanLocation = findMeanLocation(meanLocation, location);
     }
 
 
     /**
-     * Adds a new cluster (with the specified location) into the cluster list.
-     * @param location the Location with which the new cluster is to be initialized.
+     * Returns the number of elements in the cluster.
+     * @return the number of elements.
      */
-    private void addCluster(Location location) {
-        PseudoClusterElement element = new PseudoClusterElement(location);
-        clusterList.add(element);
+    int size() {
+        return locationCluster.size();
     }
 
 
     /**
-     * Finds the cluster with the largest number of elements.
-     * @return the largest cluster.
+     * Returns the list of locations currently part of the cluster.
+     * @return the list of cluster locations.
      */
-    PseudoClusterElement getLargestCluster() {
-        int size = clusterList.size();
-        if (size == 0) {
-            return null;
-        }
-        else if (size == 1) {
-            return clusterList.get(0);
-        }
-        else {
-            int maxIndex = 0;
-            int maxSize = clusterList.get(0).size();
+    List<Location> getLocations() {
+        return locationCluster;
+    }
 
-            for (int i = 1; i < size; i++) {
-                int tempSize = clusterList.get(i).size();
-                if (tempSize > maxSize) {
-                    maxIndex = i;
-                    maxSize = tempSize;
-                }
+
+    /**
+     * Returns the mean of locations in the cluster (already calculated).
+     * @return the Location representing the mean of all current cluster locations.
+     */
+    Location getMeanLocation() {
+        return meanLocation;
+    }
+
+
+    /**
+     * Finds a mean of two locations.
+     * @param location1 the first location.
+     * @param location2 the first location.
+     * @return the mean of the two locations.
+     */
+    private static Location findMeanLocation(Location location1, Location location2) {
+        Location location = new Location("");
+        location.setLatitude((location1.getLatitude() + location2.getLatitude()) / 2.0);
+        location.setLongitude((location1.getLongitude() + location2.getLongitude()) / 2.0);
+        return location;
+    }
+
+
+    /**
+     * Returns all locations in the current cluster in string format.
+     * @return a string representation of all cluster locations.
+     */
+    String locationsToString() {
+        int size = locationCluster.size();
+        if (size == 0) return "";
+        else {
+            StringBuilder sb = new StringBuilder();
+            for(Location location : locationCluster) {
+                sb.append(location.getLatitude())
+                        .append(" ")
+                        .append(location.getLongitude())
+                        .append("\n");
             }
-            return clusterList.get(maxIndex);
+            return sb.toString();
         }
     }
+
 }
